refactor: restructure codebase
