+++ src/components/team.jsx
@@ -1,7 +1,7 @@
-import con1 from '../images/4.png'
-import con2 from '../images/1.png'
-import con3 from '../images/3.png'
-import con4 from '../images/coworking.png'
+import con1 from '../assets/4.png'
+import con2 from '../assets/1.png'
+import con3 from '../assets/3.png'
+import con4 from '../assets/coworking.png'
 import { useNavigate } from 'react-router-dom'
 
 const Team = () => {
