+++ src/pages/About.jsx
@@ -1,7 +1,7 @@
-import banner1 from '../images/designer-desk-2 1.png'
-import banner2 from '../images/business-presentation-1 1.png'
-import Team from './team'
-import Header from './header'
+import banner1 from '../assets/designer-desk-2.png'
+import banner2 from '../assets/business-presentation-1.png'
+import Team from '../components/Team'
+import Header from '../components/Navbar'
 
 const About = () => {
   return (
